# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Methodly",
        "description": "Dokumentasi API untuk sistem Backend Methodly.",
        "contact": {
            "email": "support@methodly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server API Methodly"
        }
    ],
    "paths": {
        "/api/v1/admin/permissions": {
            "get": {
                "tags": [
                    "Manajemen Izin"
                ],
                "summary": "Daftar semua izin",
                "operationId": "84e762400fbd145cfc4ec6067b3d7403",
                "responses": {
                    "200": {
                        "description": "Daftar izin berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Manajemen Izin"
                ],
                "summary": "Buat izin baru",
                "operationId": "d89fe3e43f660e5007c4d0f458689c52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "delete articles"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Izin berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/permissions/{id}": {
            "delete": {
                "tags": [
                    "Manajemen Izin"
                ],
                "summary": "Hapus izin",
                "operationId": "f024f2e22f0c6dc0b29dc90445c37f51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Izin berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/roles": {
            "get": {
                "tags": [
                    "Manajemen Role"
                ],
                "summary": "Daftar semua role",
                "operationId": "90631fac9c850d5d94147a2809e94f31",
                "responses": {
                    "200": {
                        "description": "Daftar role berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Manajemen Role"
                ],
                "summary": "Buat role baru",
                "operationId": "e87448a055011802d3f694eda07e8c1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "editor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/roles/{id}": {
            "get": {
                "tags": [
                    "Manajemen Role"
                ],
                "summary": "Detail role",
                "operationId": "0cb96e8a9a318a0655368c9f2fc1be6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail role berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Manajemen Role"
                ],
                "summary": "Update nama role",
                "operationId": "3ddd86b403c0cfedd28fdf03f7ef7d0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "editor-senior"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Manajemen Role"
                ],
                "summary": "Hapus role",
                "operationId": "7d91f04ddbc9ec3b4bda5565e82691f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/roles/{id}/permissions": {
            "post": {
                "tags": [
                    "Manajemen Role"
                ],
                "summary": "Sinkronisasi izin ke role",
                "description": "Menghapus semua izin lama dan menggantinya dengan daftar izin yang baru",
                "operationId": "52b8a17b2c0845696504d234564845b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permissions"
                                ],
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "manage users"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Izin berhasil disinkronisasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Manajemen User Admin"
                ],
                "summary": "Daftar semua user",
                "operationId": "5277e8d1c017eaf6a72441766b4d1e25",
                "responses": {
                    "200": {
                        "description": "Daftar user berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}/roles": {
            "post": {
                "tags": [
                    "Manajemen User Admin"
                ],
                "summary": "Sinkronisasi role ke user",
                "operationId": "5eafba5e99c42559c33af1ad3dcdf60a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roles"
                                ],
                                "properties": {
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "instructor"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role berhasil diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}": {
            "delete": {
                "tags": [
                    "Manajemen User Admin"
                ],
                "summary": "Hapus user",
                "operationId": "7fa62a51fcbadc484b351d7c931a474c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer untuk akses API",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}